+++ tests/fixtures/noqa/noqa.py
@@ -722,3 +722,5 @@ try:
 except ValueError or TypeError:  # noqa: WPS455
     my_print("Oops.")
 
+if float("NaN") < number:  # noqa: WPS456
+    my_print("Greater than... what?")