+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Semantic versioning in our case means:
 
 - Forbids using non-trivial expressions as an argument to `except`
 - Forbids using too many variables in a tuple unpacking
+- Forbid using `float("NaN")`.
 - Allow `__call__` method to be asynchronous
 - Allows common strings not to be counted against string constant overuse limit